requirements.txt
# Core ML dependencies
tensorflow>=2.13.0,<2.18.0
keras>=3.0.0
numpy>=1.23.0
opencv-python>=4.8.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Streamlit (existing)
streamlit>=1.28.0

# Data science
scikit-learn>=1.3.0
matplotlib>=3.7.0
Pillow>=10.0.0
requests>=2.31.0
tabulate>=0.9.0